More tests to ImageCaptioningEngine.
